chore: Temporarily disable iOS CI build
To work around persistent signing issues, the `build-ios` job is temporarily disabled in the GitHub Actions workflow.

The workflow will now only build the Android app and upload the AAB to a draft release.

The Fastfile has also been updated with an `archive_only` lane to prepare for a manual signing process for iOS builds in the future.