feat(video): Use direct video playback for attachments
Displaying video thumbnails using the Skia librarys `useVideo` hook was unreliable, possibly due to an underlying issue in the library.

This change replaces the thumbnailing attempt with direct video playback within the Skia component. It also includes an optimization to only load and play the video when it is set to be visible, conserving resources.

Closes #59